Change value of BasicAuthenticationInterceptor.PRINCIPAL_KEY
This is a breaking change for applications that are using this
library with basic authentication. The change needed in those
applications is to change the key requested from the message
context to the new value of PRINCIPAL_KEY, which is now
"dropwizard.jakarta.xml.ws.principal"

Also, added the current date/time in the output of
JavaFirstServiceImpl and the two HTTP resource classes, so that
it is clear that new requests are being made, e.g. in a browser

Closes #74