Rename internal fields, vars, etc. so they don't include "jaxws"
Rename fields, local vars, parameters, etc. so they don't include
"jaxws" in them. For now, I have chosen to use "jws" to stand
for Jakarta XML Web Services, which I think is easier to read than
"jxws", e.g. jwsEnvironment

Fix issue in README that referred tot eh old "jaxws" directory
in the section on building an uber JAR.

Also, rename dropwizard-jaxws-changelog.md to
legacy-dropwizard-jaxws-changelog.md, mainly so that it is more
clear that it is a historical artifact.

Closes #77