Code cleanup: UnitOfWorkInvoker and UnitOfWorkInvokerFactory (#98)
* Update obsolete javadoc comment in UnitOfWorkInvoker. Dropwizard made
UnitOfWorkAspect public a long time ago, in version 1.1.0.
* Modify UnitOfWorkInvoker#invoke to use try-with-resources for the
Hibernate Session, and wrap the Map#get call with requireNonNull since
the containsKey check returned true. Remove redundant RuntimeException
type parameter in call to rethrow.
* Use var in UnitOfWorkInvokerFactory for the ImmutableMap.Builder to
reduce code verbosity
* Refactor logic in UnitOfWorkInvokerFactory so that it doesn't reassign
a local variable. Instead, if the "unitOfWorkMethods" map is empty,
return the rootInvoker. This inverts the logic and is clearer, since
re-assigning variables is generally not a good practice.